[Kernel] Change uint32_t to uintptr_t

+++ Kernel/Core/GDT.cpp
@@ -5,13 +5,13 @@
 GDT GDT::m_Instance;
 
 GDT::GDT()
-    : m_Entries(), m_Ptr({ m_ActiveEntries, (uint32_t)m_Entries })
+    : m_Entries(), m_Ptr({ m_ActiveEntries, (uintptr_t)m_Entries })
 {
 }
 
 void GDT::Install()
 {
-    m_Ptr.Address = (uint32_t)m_Entries;
+    m_Ptr.Address = (uintptr_t)m_Entries;
 
     asm volatile("lgdt %0\n"
         "mov %1, %%ds\n"
