+++ Kernel/Core/GDT.h
@@ -56,7 +56,7 @@ private:
     struct PACKED Ptr
     {
         uint16_t Size;
-        uint32_t Address;
+        uint64_t Address;
     } m_Ptr;
 
     static GDT m_Instance;
